1.16.5

+++ src/main/java/zmaster587/advancedRocketry/inventory/modules/ModulePlanetImage.java
@@ -7,14 +7,11 @@ import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.inventory.container.Container;
 import net.minecraft.util.math.vector.Quaternion;
 
-import org.lwjgl.opengl.GL11;
-
 import com.mojang.blaze3d.matrix.MatrixStack;
 
 import zmaster587.advancedRocketry.api.Constants;
 import zmaster587.advancedRocketry.client.render.planet.RenderPlanetarySky;
 import zmaster587.advancedRocketry.dimension.DimensionProperties;
-import zmaster587.advancedRocketry.stations.SpaceObjectManager;
 import zmaster587.libVulpes.inventory.modules.ModuleBase;
 
 public class ModulePlanetImage extends ModuleBase {
@@ -44,7 +41,7 @@ public class ModulePlanetImage extends ModuleBase {
 		//GL11.glTranslatef(xPosition, 100 + this.zLevel, yPosition);
 		float newWidth = width/2f;
 
-		RenderPlanetarySky.renderPlanetPubHelper(vertexbuffer, matrix, properties.getPlanetIcon(), (int)(x + this.offsetX + newWidth), (int)(y + this.offsetY + newWidth), (double)-0.1, newWidth, 1f, properties.getSolarTheta(), properties.hasAtmosphere(), properties.skyColor, properties.ringColor, properties.isGasGiant(), properties.hasRings(), properties.hasDecorators(), new float[]{0, 0, 0}, 1f);
+		RenderPlanetarySky.renderPlanetPubHelper(vertexbuffer, matrix, properties.getPlanetIcon(), (int)(x + this.offsetX + newWidth), (int)(y + this.offsetY + newWidth), -0.1, newWidth, 1f, properties.getSolarTheta(), properties.hasAtmosphere(), properties.skyColor, properties.ringColor, properties.isGasGiant(), properties.hasRings(), properties.hasDecorators(), new float[]{0, 0, 0}, 1f);
 		matrix.pop();
 	}
